Fixed u0 so that orbits have the fixed mass always at a foci. Moved the test coded to test.py and emptied main.py for main code.
